requirements.txt
# requirements.txt
requestsFlask
requests
Flask-CORS
gunicorn